requirements.txt
streamlit==1.40.0
langchain==0.3.4
langchain-core==0.3.12
langchain-community==0.3.2
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.0
huggingface-hub==0.26.1
sentence-transformers==3.0.1
faiss-cpu==1.11.0
python-dotenv==1.0.1
PyPDF2==3.0.1
